chore(bench): AC-W2's dispatch budget leaves the merge gate
A main-health run read p50=13.19ms against the 200ms budget — sixteen times
inside it — and p95=201.63ms, and filed main as broken over the 1.63ms. Six
integration shards share one Postgres on that runner, so the eleventh-largest
of two hundred samples is a measurement of the runner rather than of the
dispatch. Raising the sample count from 25 to 200 had already been tried and
bought exactly one run.

This was the last clock left in the lane. PERF-3/PERF-7, PERF-1/PERF-4 and
CAP-PARAM-1 sit behind the bench tag; MOBILE-AC-2 sits behind bench-mobile;
and ci.yml states the rule on the uat lane in as many words — nothing in it
reads a clock, so a slow runner cannot redden it. Run by hand the budget
passes with room to spare, p95=25.7ms, which is the whole argument.

Nothing replaces it in the standing lane, and the file says so rather than
implying cover it does not have. PERF-3/PERF-7 could leave because a
deterministic test already held its regression; AC-W2 has no sibling, and
three candidate instruments were written and measured before this landed —
a seq_scan delta, the same over history seeded under the firing handler, and
a rows-read ceiling. None can fail: the engine dispatches over the app pool
and pg_stat_force_next_flush reaches only the calling backend, so a reader on
the owner connection sees counters the dispatch has not reported yet. A gate
that cannot fail is worse than an absent one.
